Revert "Some name updates"
This reverts commit d39c90525fe5eb2c4c93903c96be0bcf1a6f2a73.

+++ src/main/java/com/flansmod/common/teams/EntityFlag.java
@@ -221,7 +221,7 @@ public class EntityFlag extends Entity implements ITeamObject
 	
 	public int getTeamID()
 	{
-		return (int)getDataManager().get(TEAMID);
+		return getDataManager().get(TEAMID);
 	}
 	
 	@Override
